+++ packages/bugsnag_flutter_performance/lib/src/uploader/sampling_probability_store.dart
@@ -10,6 +10,7 @@ abstract class SamplingProbabilityStore {
     double samplingProbability,
     DateTime expireDate,
   );
+  void setFixedProbability(double? samplingProbability);
 }
 
 class SamplingProbabilityStoreImpl implements SamplingProbabilityStore {
@@ -19,13 +20,19 @@ class SamplingProbabilityStoreImpl implements SamplingProbabilityStore {
   final BugsnagClock clock;
 
   double? _samplingProbability;
+  double? _fixedSamplingProbability;
   DateTime? _expireDate;
   var _isInitialized = false;
 
-  SamplingProbabilityStoreImpl(this.clock);
+  SamplingProbabilityStoreImpl(
+    this.clock,
+  );
 
   @override
   Future<double?> get samplingProbability async {
+    if (_fixedSamplingProbability != null) {
+      return _fixedSamplingProbability!;
+    }
     await _initializeIfNeeded();
     if (_expireDate != null && _expireDate!.isBefore(clock.now())) {
       return null;
@@ -38,6 +45,9 @@ class SamplingProbabilityStoreImpl implements SamplingProbabilityStore {
     double samplingProbability,
     DateTime expireDate,
   ) async {
+    if (_fixedSamplingProbability != null) {
+      return;
+    }
     await _initializeIfNeeded();
     if (_samplingProbability != null &&
         _samplingProbability! < samplingProbability) {
@@ -71,6 +81,11 @@ class SamplingProbabilityStoreImpl implements SamplingProbabilityStore {
     }
 
     _isInitialized = true;
+  }
+
+  @override
+  void setFixedProbability(double? samplingProbability) {
+    _fixedSamplingProbability = samplingProbability;
   }
 
   Future<void> _writeToFile(